creo y refactorizo en la clase controlador.
